Add files via upload
In machine learning, splitting data into training and test sets is known as resampling, or more generally as cross validation. This is an important step in the model fitting process because it allows us to estimate how our trained machine learning algorithms with perform on new data. 

The ultimate goal for any machine learning algorithm is to provide accurate predictions on new, previously unseen data.

Resampling is achieved with the `rsample` package from `tidymodels`. To demonstrate how this done, let's import the `tidymodels` package and the `employee_data`.

The `tidymodels` package loads the core machine learning packages that we will be using this semester, including `rsample`, `recipes`, `parsnip`, `yardstick`, `dials`, `tune`, and `workflows`. Each one of these packages serves a specific role in the modeling process. This tutorial will focus on resampling with `rsample` and feature engineering with `recipes`.

We also load the `tidyverse` package for reading in our data and data manipulation.